fix: clippy + file-size cap for post-mortem PR (#120)
Three CI failures on dev:

1. Clippy `too_many_arguments` error on `AuditRing::push_ts` in
   rs-delivery/src/audit_ring.rs (7/6 args). Add `#[allow(..)]` to the
   helper — signature is intentional (mirrors push() + ts override).

2. `crates/rs-api/src/delivery.rs` was 1358 lines (cap 1000). Split:
   - `delivery_status.rs` (new) — `EndpointRestartRecord`,
     `EndpointDeliveryStatus`, `DeliveryStatus`, `YouTubeStatus`,
     `pick_last_error_line_inline`, `load_restart_history_from_db`,
     plus the `get_delivery_status` + `poll_delivery_metrics` impl
     methods.
   - `delivery_audit_mirror.rs` (new) — `mirror_vps_audit` free function.
   - `delivery.rs` keeps `DeliveryOrchestrator` struct + lifecycle
     methods (`new`, `start_delivery`, `poll_and_init`, `stop_delivery`,
     `monitor_delivery_health`, `find_delivery_image`).
   - Re-exports preserved at `crate::delivery::*` so existing callers
     (delivery_youtube, delivery_tests, delivery_handlers, lib.rs)
     compile unchanged.
   - Added pub(crate) `endpoint_fast_cache_lock` accessor so the
     sibling `delivery_status.rs` can read the cache.

3. `crates/rs-core/src/db/tests.rs` was 1005 lines (5 over cap).
   Extracted the two pool PRAGMA tests into `db/pool_tests.rs`,
   registered in `db/mod.rs`.

Final line counts:
 - audit_ring.rs: 190
 - delivery.rs: 933
 - delivery_status.rs: 371
 - delivery_audit_mirror.rs: 107
 - tests.rs: 978
 - pool_tests.rs: 31
