+++ crates/rs-api/src/lib.rs
@@ -1,8 +1,10 @@
 pub mod audit_handlers;
 pub mod delivery;
+pub(crate) mod delivery_audit_mirror;
 pub mod delivery_endpoints;
 pub mod delivery_handlers;
 pub(crate) mod delivery_helpers;
+pub(crate) mod delivery_status;
 #[cfg(test)]
 mod delivery_tests;
 mod delivery_youtube;
