+++ crates/rs-core/src/db/tests.rs
@@ -975,31 +975,4 @@ async fn update_event_rescue_video_url() {
 // Delivery log capture tests are in delivery_log_tests.rs
 // Upload telemetry tests are in upload_tests.rs
 // Migration idempotency tests are in migration_tests.rs
-
-#[tokio::test]
-async fn create_pool_sets_busy_timeout_and_synchronous() {
-    let tmp = tempfile::NamedTempFile::new().unwrap();
-    let pool = crate::db::create_pool(tmp.path()).await.unwrap();
-
-    let busy_timeout: i64 = sqlx::query_scalar("PRAGMA busy_timeout")
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-    assert_eq!(busy_timeout, 5000, "busy_timeout must be 5000 ms");
-
-    let sync: i64 = sqlx::query_scalar("PRAGMA synchronous")
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-    assert_eq!(sync, 1, "synchronous must be NORMAL (1), got {sync}");
-}
-
-#[tokio::test]
-async fn create_memory_pool_sets_busy_timeout() {
-    let pool = crate::db::create_memory_pool().await.unwrap();
-    let busy_timeout: i64 = sqlx::query_scalar("PRAGMA busy_timeout")
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-    assert_eq!(busy_timeout, 5000);
-}
+// Pool PRAGMA tests are in pool_tests.rs
